requirements.txt
jupyterlab
datasets
pillow
matplotlib
ultralytics
mlflow
minio
bentoml
pydantic